refactorings for supporting custom endpoints with examples
