+++ exec/src/main/java/com/google/fhir/gateway/MainApp.java
@@ -19,6 +19,10 @@ import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.boot.web.servlet.ServletComponentScan;
 
+/**
+ * This class shows the minimum that is required to create a FHIR Gateway with all AccessChecker
+ * plugins defined in "com.google.fhir.gateway.plugin".
+ */
 @SpringBootApplication(scanBasePackages = {"com.google.fhir.gateway.plugin"})
 @ServletComponentScan(basePackages = "com.google.fhir.gateway")
 public class MainApp {
