+++ server/src/main/java/com/google/fhir/gateway/HttpFhirClient.java
@@ -37,6 +37,8 @@ import org.apache.http.impl.client.HttpClients;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
+// TODO evaluate if we can provide the API of HAPI's IGenericClient as well:
+//  https://hapifhir.io/hapi-fhir/docs/client/generic_client.html
 public abstract class HttpFhirClient {
 
   private static final Logger logger = LoggerFactory.getLogger(HttpFhirClient.class);
@@ -102,6 +104,7 @@ public abstract class HttpFhirClient {
     }
   }
 
+  /** This method is intended to be used only for requests that are relayed to the FHIR store. */
   HttpResponse handleRequest(ServletRequestDetails request) throws IOException {
     String httpMethod = request.getServletRequest().getMethod();
     RequestBuilder builder = RequestBuilder.create(httpMethod);
